knowledge_graph/requirements.txt
# 知识图谱系统依赖
neo4j==5.15.0
pandas==2.0.3
numpy==1.24.3
flask==2.3.3
flask-cors==4.0.0
networkx==3.1
plotly==5.17.0
dash==2.14.1
dash-cytoscape==0.3.0
python-dotenv==1.0.0
requests==2.31.0 